+++ trunk/src/org/simbrain/world/visionworld/pixelmatrix/ScreenCapturePixelMatrix.java
@@ -33,8 +33,7 @@ import org.simbrain.world.visionworld.ReceptiveField;
 /**
  * Screen capture pixel matrix.
  */
-public final class ScreenCapturePixelMatrix
-    implements PixelMatrix {
+public final class ScreenCapturePixelMatrix implements PixelMatrix {
 
     /** Image for this pixel matrix. */
     private volatile BufferedImage image;
@@ -60,7 +59,6 @@ public final class ScreenCapturePixelMatrix
     /** Default width. */
     public static final int DEFAULT_WIDTH = 100;
 
-
     /**
      * Create a new screen capture pixel matrix at the specified dimensions.
      *
@@ -69,17 +67,20 @@ public final class ScreenCapturePixelMatrix
      * @param width width in pixels, must be &gt; 0
      * @param height height in pixels, must be &gt; 0
      */
-    public ScreenCapturePixelMatrix(final int originX, final int originY, final int width, final int height) {
+    public ScreenCapturePixelMatrix(final int originX, final int originY,
+            final int width, final int height) {
         if (height <= 0) {
-            throw new IllegalArgumentException("height must be greater than zero");
+            throw new IllegalArgumentException(
+                    "height must be greater than zero");
         }
         if (width <= 0) {
-            throw new IllegalArgumentException("width must be greater than zero");
+            throw new IllegalArgumentException(
+                    "width must be greater than zero");
         }
-        this.image = new BufferedImage(width, height, BufferedImage.TYPE_INT_ARGB);
+        this.image = new BufferedImage(width, height,
+                BufferedImage.TYPE_INT_ARGB);
         this.propertyChangeSupport = new PropertyChangeSupport(this);
     }
-
 
     /**
      * Throw an ArrayIndexOutOfBoundsException if either of the specified
@@ -87,21 +88,27 @@ public final class ScreenCapturePixelMatrix
      *
      * @param x x coordinate to check
      * @param y y coordinate to check
-     * @throws ArrayIndexOutOfBoundsException if either of the specified coordinates
-     *    are outside the bounds of this pixel matrix
+     * @throws ArrayIndexOutOfBoundsException if either of the specified
+     *             coordinates are outside the bounds of this pixel matrix
      */
     private void checkCoordinates(final int x, final int y) {
         if (x < 0) {
-            throw new ArrayIndexOutOfBoundsException("x must be greater than 0, was " + x);
+            throw new ArrayIndexOutOfBoundsException(
+                    "x must be greater than 0, was " + x);
         }
         if (y < 0) {
-            throw new ArrayIndexOutOfBoundsException("y must be greater than 0, was " + y);
+            throw new ArrayIndexOutOfBoundsException(
+                    "y must be greater than 0, was " + y);
         }
-        if (x > (int) (getWidth() - 1)) {
-            throw new ArrayIndexOutOfBoundsException("x must be less than or equal to (getWidth() - 1), was " + x);
+        if (x > (getWidth() - 1)) {
+            throw new ArrayIndexOutOfBoundsException(
+                    "x must be less than or equal to (getWidth() - 1), was "
+                            + x);
         }
-        if (y > (int) (getHeight() - 1)) {
-            throw new ArrayIndexOutOfBoundsException("y must be less than or equal to (getHeight() - 1), was " + y);
+        if (y > (getHeight() - 1)) {
+            throw new ArrayIndexOutOfBoundsException(
+                    "y must be less than or equal to (getHeight() - 1), was "
+                            + y);
         }
     }
 
@@ -125,18 +132,18 @@ public final class ScreenCapturePixelMatrix
      */
     public void capture() {
         BufferedImage oldImage = this.image;
-        try
-        {
-            // TODO:  or draw into existing image?
-            Rectangle rect = new Rectangle(originX, originY, getWidth(), getHeight());
+        try {
+            // TODO: or draw into existing image?
+            Rectangle rect = new Rectangle(originX, originY, getWidth(),
+                    getHeight());
             this.image = new Robot().createScreenCapture(rect);
-            propertyChangeSupport.firePropertyChange("image", oldImage, this.image);
-        }
-        catch (AWTException e) {
+            propertyChangeSupport.firePropertyChange("image", oldImage,
+                    this.image);
+        } catch (AWTException e) {
             e.printStackTrace();
         }
     }
-    
+
     /** {@inheritDoc} */
     public Color getPixel(final int x, final int y) {
         checkCoordinates(x, y);
@@ -157,7 +164,8 @@ public final class ScreenCapturePixelMatrix
         if (color == null) {
             throw new IllegalArgumentException("color must not be null");
         }
-        int rgb = (color.getRed() << 16) | (color.getGreen() << 8) | (color.getBlue());
+        int rgb = (color.getRed() << 16) | (color.getGreen() << 8)
+                | (color.getBlue());
         int[] a = new int[1];
         a[0] = color.getAlpha();
         image.setRGB(x, y, rgb);
@@ -173,27 +181,30 @@ public final class ScreenCapturePixelMatrix
 
     /** {@inheritDoc} */
     public void addPropertyChangeListener(final String propertyName,
-                                          final PropertyChangeListener listener) {
+            final PropertyChangeListener listener) {
         propertyChangeSupport.addPropertyChangeListener(propertyName, listener);
     }
 
     /** {@inheritDoc} */
-    public void removePropertyChangeListener(final PropertyChangeListener listener) {
+    public void removePropertyChangeListener(
+            final PropertyChangeListener listener) {
         propertyChangeSupport.removePropertyChangeListener(listener);
     }
 
     /** {@inheritDoc} */
     public void removePropertyChangeListener(final String propertyName,
-                                             final PropertyChangeListener listener) {
-        propertyChangeSupport.removePropertyChangeListener(propertyName, listener);
+            final PropertyChangeListener listener) {
+        propertyChangeSupport.removePropertyChangeListener(propertyName,
+                listener);
     }
 
     /** {@inheritDoc} */
     public Image view(final ReceptiveField receptiveField) {
         if (receptiveField == null) {
-            throw new IllegalArgumentException("receptiveField must not be null");
+            throw new IllegalArgumentException(
+                    "receptiveField must not be null");
         }
         return image.getSubimage(receptiveField.getX(), receptiveField.getY(),
-                                 receptiveField.getWidth(), receptiveField.getHeight());
+                receptiveField.getWidth(), receptiveField.getHeight());
     }
 }
